Implement RecognizersRunnerListener
