+++ app/src/main/java/io/github/driverassistant/RecognizersRunner.kt
@@ -2,19 +2,16 @@ package io.github.driverassistant
 
 import android.os.HandlerThread
 import android.util.Log
-import android.widget.TextView
 import io.github.driverassistant.recognizer.LatestImage
 import io.github.driverassistant.recognizer.Recognizer
 import io.github.driverassistant.util.camera.ShootingCamera
 import io.github.driverassistant.util.handler
-import io.github.driverassistant.util.postApply
 import io.github.driverassistant.util.postDelayed
 import kotlin.math.roundToLong
 
 class RecognizersRunner(
     fps: Double,
-    private val statsTextView: TextView,
-    private val recognizedObjectsView: RecognizedObjectsView,
+    private val recognizersRunnerListener: RecognizersRunnerListener,
     private val recognizers: Iterable<Recognizer>,
     private val shootingCamera: ShootingCamera,
     private val captureThread: HandlerThread
@@ -26,21 +23,21 @@ class RecognizersRunner(
 
         captureThread.handler.postDelayed(delay) { shootingCamera.submitRequestForNextShot() }
 
-        val statsText = with(latestImage) { "${bytes.size} bytes, $width x $height" }
-        statsTextView.postApply { text = statsText }
-
-        val paintablesList = recognizers
+        val paintables = recognizers
             .flatMap { recognizer -> recognizer.recognize(latestImage) }
             .flatMap { recognizedObject -> recognizedObject.elements }
             .map { recognizedObjectElement -> recognizedObjectElement.toPaintableOnCanvas() }
 
-        recognizedObjectsView.postApply {
-            paintables = paintablesList
-            invalidate()
-        }
+        recognizersRunnerListener.onResult(latestImage, paintables)
     }
 
     companion object {
         private const val TAG = "RecognizersRunner"
+
+        interface RecognizersRunnerListener {
+            fun onResult(latestImage: LatestImage, objects: Iterable<PaintableOnCanvas>)
+
+            fun onEnd()
+        }
     }
 }
