--- app/src/main/java/io/github/driverassistant/state/common/AfterRecognizerOperations.kt
@@ -1,16 +1,0 @@
-package io.github.driverassistant.state.common
-
-import android.widget.TextView
-import io.github.driverassistant.RecognizedObjectsView
-import io.github.driverassistant.util.postApply
-
-fun cleanRecognizersScreen(
-    recognizedObjectsView: RecognizedObjectsView,
-    statsTextView: TextView
-) {
-    recognizedObjectsView.postApply {
-        paintables = emptyList()
-        invalidate()
-    }
-    statsTextView.postApply { text = "" }
-}
